+++ packages/openfort-react/src/constants/openfort.ts
@@ -1,2 +1,1 @@
 export const embeddedWalletId = 'xyz.openfort';
-export const emailToVerifyLocalStorageKey = "openfort:email-to-verify";